refactor: Move unit tests in relayer.rs into separate file (#6690)
